poops and slovar dvoretskogo + lots of minor changes (probably)

+++ src/main/java/net/svisvi/jigsaw/procedures/LivingBeaverEntityIsHurtProcedure.java
@@ -101,9 +101,9 @@ public class LivingBeaverEntityIsHurtProcedure {
 					}
 				}
 				((_spawnentity instanceof LivingEntity _entGetArmor ? _entGetArmor.getItemBySlot(EquipmentSlot.HEAD) : ItemStack.EMPTY))
-						.setHoverName(new TextComponent((((sourceentity instanceof LivingEntity _livEnt ? _livEnt.getMainHandItem() : ItemStack.EMPTY).getDisplayName().getString()).substring(1,
+						.setHoverName(new TextComponent((((sourceentity instanceof LivingEntity _livEnt ? _livEnt.getMainHandItem() : ItemStack.EMPTY).getDisplayName().getString()).substring((int) 1,
 								(int) (((sourceentity instanceof LivingEntity _livEnt ? _livEnt.getMainHandItem() : ItemStack.EMPTY).getDisplayName().getString()).length() - 1)))));
-				_spawnentity.setCustomName(new TextComponent((((sourceentity instanceof LivingEntity _livEnt ? _livEnt.getMainHandItem() : ItemStack.EMPTY).getDisplayName().getString()).substring(1,
+				_spawnentity.setCustomName(new TextComponent((((sourceentity instanceof LivingEntity _livEnt ? _livEnt.getMainHandItem() : ItemStack.EMPTY).getDisplayName().getString()).substring((int) 1,
 						(int) (((sourceentity instanceof LivingEntity _livEnt ? _livEnt.getMainHandItem() : ItemStack.EMPTY).getDisplayName().getString()).length() - 1)))));
 				if (_spawnentity instanceof Mob _spawnmob)
 					_spawnmob.finalizeSpawn(_serverLevel, world.getCurrentDifficultyAt(_spawnentity.blockPosition()), MobSpawnType.MOB_SUMMONED, null, null);
